refactor: move gateway to its own package
making room for an rpc provider
