+++ account.go
@@ -4,15 +4,20 @@ import (
 	"context"
 	"fmt"
 	"math/big"
-	"net/http"
+)
+
+const (
+	EXECUTE_SELECTOR    string = "__execute__"
+	TRANSACTION_PREFIX  string = "invoke"
+	TRANSACTION_VERSION int64  = 0
 )
 
 type Signer struct {
-	private *big.Int
-	Curve   StarkCurve
-	Gateway *StarknetGateway
-	PublicX *big.Int
-	PublicY *big.Int
+	private  *big.Int
+	Curve    StarkCurve
+	provider Provider
+	PublicX  *big.Int
+	PublicY  *big.Int
 }
 
 type FeeEstimate struct {
@@ -27,17 +32,17 @@ type FeeEstimate struct {
 	- full StarknetGateway definition
 	- public key pair for signature verifications
 */
-func (sc StarkCurve) NewSigner(private, pubX, pubY *big.Int, gw *StarknetGateway) (*Signer, error) {
+func (sc StarkCurve) NewSigner(private, pubX, pubY *big.Int, provider Provider) (*Signer, error) {
 	if len(sc.ConstantPoints) == 0 {
 		return nil, fmt.Errorf("must initiate precomputed constant points")
 	}
 
 	return &Signer{
-		private: private,
-		Curve:   sc,
-		Gateway: gw,
-		PublicX: pubX,
-		PublicY: pubY,
+		private:  private,
+		Curve:    sc,
+		provider: provider,
+		PublicX:  pubX,
+		PublicY:  pubY,
 	}, nil
 }
 
@@ -46,22 +51,26 @@ func (sc StarkCurve) NewSigner(private, pubX, pubY *big.Int, gw *StarknetGateway
 	- implementation has been tested against OpenZeppelin Account contract as of: https://github.com/OpenZeppelin/cairo-contracts/blob/4116c1ecbed9f821a2aa714c993a35c1682c946e/src/openzeppelin/account/Account.cairo
 	- accepts a multicall
 */
-func (signer *Signer) Execute(ctx context.Context, address string, txs []Transaction) (addResp AddTxResponse, err error) {
-	nonce, err := signer.Gateway.AccountNonce(ctx, address)
+func (signer *Signer) Execute(ctx context.Context, address string, txs []Transaction) (AddTxResponse, error) {
+	nonce, err := signer.provider.AccountNonce(ctx, address)
 	if err != nil {
-		return addResp, err
+		return AddTxResponse{}, err
 	}
 
 	maxFee := big.NewInt(0)
-
-	hash, err := signer.Curve.HashMulticall(address, nonce, maxFee, UTF8StrToBig(signer.Gateway.ChainId), txs)
+	chainID, err := signer.provider.ChainID(ctx)
 	if err != nil {
-		return addResp, err
+		return AddTxResponse{}, err
+	}
+
+	hash, err := signer.Curve.HashMulticall(address, nonce, maxFee, UTF8StrToBig(chainID), txs)
+	if err != nil {
+		return AddTxResponse{}, err
 	}
 
 	r, s, err := signer.Curve.Sign(hash, signer.private)
 	if err != nil {
-		return addResp, err
+		return AddTxResponse{}, err
 	}
 
 	req := Transaction{
@@ -71,31 +80,7 @@ func (signer *Signer) Execute(ctx context.Context, address string, txs []Transac
 		Signature:          []string{r.String(), s.String()},
 	}
 
-	return signer.Gateway.Invoke(ctx, req)
-}
-
-func (sg *StarknetGateway) AccountNonce(ctx context.Context, address string) (nonce *big.Int, err error) {
-	resp, err := sg.Call(ctx, Transaction{
-		ContractAddress:    address,
-		EntryPointSelector: "get_nonce",
-	}, nil)
-	if err != nil {
-		return nonce, err
-	}
-	if len(resp) == 0 {
-		return nonce, fmt.Errorf("no resp in contract call 'get_nonce' %v", address)
-	}
-
-	return HexToBN(resp[0]), nil
-}
-
-func (sg *StarknetGateway) EstimateFee(ctx context.Context, tx Transaction) (fee FeeEstimate, err error) {
-	req, err := sg.newRequest(ctx, http.MethodPost, "/estimate_fee", tx)
-	if err != nil {
-		return fee, err
-	}
-
-	return fee, sg.do(req, &fee)
+	return signer.provider.Invoke(ctx, req)
 }
 
 func (sc StarkCurve) HashMulticall(addr string, nonce, maxFee, chainId *big.Int, txs []Transaction) (hash *big.Int, err error) {
