+++ gateway/storage_test.go
@@ -1,4 +1,4 @@
-package caigo
+package gateway
 
 import (
 	"context"
