+++ gateway/transaction.go
@@ -1,4 +1,4 @@
-package caigo
+package gateway
 
 import (
 	"context"
@@ -9,80 +9,9 @@ import (
 	"strings"
 	"time"
 
+	"github.com/dontpanicdao/caigo"
 	"github.com/google/go-querystring/query"
 )
-
-/*
-	StarkNet transaction states
-*/
-const (
-	NOT_RECIEVED = TxStatus(iota)
-	REJECTED
-	RECEIVED
-	PENDING
-	ACCEPTED_ON_L2
-	ACCEPTED_ON_L1
-)
-
-var statuses = []string{"NOT_RECEIVED", "REJECTED", "RECEIVED", "PENDING", "ACCEPTED_ON_L2", "ACCEPTED_ON_L1"}
-
-type TxStatus int
-
-type TransactionStatus struct {
-	TxStatus        string `json:"tx_status"`
-	BlockHash       string `json:"block_hash"`
-	TxFailureReason struct {
-		ErrorMessage string `json:"error_message,omitempty"`
-	} `json:"tx_failure_reason,omitempty"`
-}
-
-// Starknet transaction composition
-type Transaction struct {
-	ContractAddress    string   `json:"contract_address"`
-	EntryPointSelector string   `json:"entry_point_selector"`
-	Calldata           []string `json:"calldata"`
-	Signature          []string `json:"signature"`
-	EntryPointType     string   `json:"entry_point_type,omitempty"`
-	TransactionHash    string   `json:"transaction_hash,omitempty"`
-	Type               string   `json:"type,omitempty"`
-	Nonce              string   `json:"nonce,omitempty"`
-}
-
-type StarknetTransaction struct {
-	TransactionIndex int         `json:"transaction_index"`
-	BlockNumber      int         `json:"block_number"`
-	Transaction      Transaction `json:"transaction"`
-	BlockHash        string      `json:"block_hash"`
-	Status           string      `json:"status"`
-}
-
-type TransactionReceipt struct {
-	Status                string `json:"status"`
-	BlockHash             string `json:"block_hash"`
-	BlockNumber           int    `json:"block_number"`
-	TransactionIndex      int    `json:"transaction_index"`
-	TransactionHash       string `json:"transaction_hash"`
-	L1ToL2ConsumedMessage struct {
-		FromAddress string   `json:"from_address"`
-		ToAddress   string   `json:"to_address"`
-		Selector    string   `json:"selector"`
-		Payload     []string `json:"payload"`
-	} `json:"l1_to_l2_consumed_message"`
-	L2ToL1Messages     []interface{} `json:"l2_to_l1_messages"`
-	Events             []interface{} `json:"events"`
-	ExecutionResources struct {
-		NSteps                 int `json:"n_steps"`
-		BuiltinInstanceCounter struct {
-			PedersenBuiltin   int `json:"pedersen_builtin"`
-			RangeCheckBuiltin int `json:"range_check_builtin"`
-			BitwiseBuiltin    int `json:"bitwise_builtin"`
-			OutputBuiltin     int `json:"output_builtin"`
-			EcdsaBuiltin      int `json:"ecdsa_builtin"`
-			EcOpBuiltin       int `json:"ec_op_builtin"`
-		} `json:"builtin_instance_counter"`
-		NMemoryHoles int `json:"n_memory_holes"`
-	} `json:"execution_resources"`
-}
 
 type TransactionOptions struct {
 	TransactionId   uint64 `url:"transactionId,omitempty"`
@@ -92,7 +21,7 @@ type TransactionOptions struct {
 // Gets the transaction information from a tx id.
 //
 // [Reference](https://github.com/starkware-libs/cairo-lang/blob/f464ec4797361b6be8989e36e02ec690e74ef285/src/starkware/starknet/services/api/feeder_gateway/feeder_gateway_client.py#L54-L58)
-func (gw *StarknetGateway) Transaction(ctx context.Context, opts TransactionOptions) (*StarknetTransaction, error) {
+func (gw *StarknetGateway) Transaction(ctx context.Context, opts TransactionOptions) (*caigo.StarknetTransaction, error) {
 	req, err := gw.newRequest(ctx, http.MethodGet, "/get_transaction", nil)
 	if err != nil {
 		return nil, err
@@ -103,7 +32,7 @@ func (gw *StarknetGateway) Transaction(ctx context.Context, opts TransactionOpti
 	}
 	appendQueryValues(req, vs)
 
-	var resp StarknetTransaction
+	var resp caigo.StarknetTransaction
 	return &resp, gw.do(req, &resp)
 }
 
@@ -115,7 +44,7 @@ type TransactionStatusOptions struct {
 // Gets the transaction status from a txn.
 //
 // [Reference](https://github.com/starkware-libs/cairo-lang/blob/fc97bdd8322a7df043c87c371634b26c15ed6cee/src/starkware/starknet/services/api/feeder_gateway/feeder_gateway_client.py#L87)
-func (gw *StarknetGateway) TransactionStatus(ctx context.Context, opts TransactionStatusOptions) (*TransactionStatus, error) {
+func (gw *StarknetGateway) TransactionStatus(ctx context.Context, opts TransactionStatusOptions) (*caigo.TransactionStatus, error) {
 	req, err := gw.newRequest(ctx, http.MethodGet, "/get_transaction_status", nil)
 	if err != nil {
 		return nil, err
@@ -126,7 +55,7 @@ func (gw *StarknetGateway) TransactionStatus(ctx context.Context, opts Transacti
 	}
 	appendQueryValues(req, vs)
 
-	var resp TransactionStatus
+	var resp caigo.TransactionStatus
 	return &resp, gw.do(req, &resp)
 }
 
@@ -171,7 +100,7 @@ func (gw *StarknetGateway) TransactionHash(ctx context.Context, id *big.Int) (st
 // Get transaction receipt for specific tx
 //
 // [Reference](https://github.com/starkware-libs/cairo-lang/blob/fc97bdd8322a7df043c87c371634b26c15ed6cee/src/starkware/starknet/services/api/feeder_gateway/feeder_gateway_client.py#L104)
-func (gw *StarknetGateway) TransactionReceipt(ctx context.Context, txHash string) (*TransactionReceipt, error) {
+func (gw *StarknetGateway) TransactionReceipt(ctx context.Context, txHash string) (*caigo.TransactionReceipt, error) {
 	req, err := gw.newRequest(ctx, http.MethodGet, "/get_transaction_receipt", nil)
 	if err != nil {
 		return nil, err
@@ -181,13 +110,13 @@ func (gw *StarknetGateway) TransactionReceipt(ctx context.Context, txHash string
 		"transactionHash": []string{txHash},
 	})
 
-	var resp TransactionReceipt
+	var resp caigo.TransactionReceipt
 	return &resp, gw.do(req, &resp)
 }
 
 // Long poll a transaction for specificed interval and max polls until the desired TxStatus has been achieved
 // or the transaction reverts
-func (gw *StarknetGateway) PollTx(ctx context.Context, txHash string, threshold TxStatus, interval, maxPoll int) (n int, status string, err error) {
+func (gw *StarknetGateway) PollTx(ctx context.Context, txHash string, threshold caigo.TxStatus, interval, maxPoll int) (n int, status string, err error) {
 	err = fmt.Errorf("could not find tx status for tx:  %s", txHash)
 
 	ticker := time.NewTicker(time.Duration(interval) * time.Second)
@@ -214,12 +143,8 @@ func (gw *StarknetGateway) PollTx(ctx context.Context, txHash string, threshold 
 	return cow, status, err
 }
 
-func (s TxStatus) String() string {
-	return statuses[s]
-}
-
 func FindTxStatus(stat string) int {
-	for i, val := range statuses {
+	for i, val := range caigo.TxStatuses {
 		if val == strings.ToUpper(stat) {
 			return i
 		}
